backend/requirements.txt
asgiref==3.4.1
Django==4.0.6
django-cors-headers==3.7.0
django-filter==22.1
django-jazzmin==2.5.0
djangorestframework==3.12.4
pytz==2021.1
sqlparse==0.4.1